# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MixPlay API",
        "description": "API da aplicação MixPlay, para pegar os dados.",
        "contact": {
            "email": "rodrigo.pinheiroa@hotmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Default Server"
        }
    ],
    "paths": {
...
        "/genre": {
            "get": {
                "tags": [
                    "Genres"
                ],
                "summary": "Listar gêneros",
                "operationId": "genreIndex",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/Genre"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#components/responses/unauthorized"
                    }
                }
            },
            "post": {
                "tags": [
                    "Genres"
                ],
                "summary": "Criar gênero",
                "operationId": "genreStore",
                "requestBody": {
                    "$ref": "#components/requestBodies/StoreGenre"
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/Genre"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#components/responses/unauthorized"
                    }
                }
            }
        },
        "/genre/{genre}": {
            "put": {
                "tags": [
                    "Genres"
                ],
                "summary": "Atualizar gênero",
                "operationId": "genreUpdate",
                "parameters": [
                    {
                        "name": "genre",
                        "in": "path",
                        "description": "ID do gênero",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#components/requestBodies/StoreGenre"
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/Genre"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#components/responses/unauthorized"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Genres"
                ],
                "summary": "Remover gênero",
                "operationId": "genreDestroy",
                "parameters": [
                    {
                        "name": "genre",
                        "in": "path",
                        "description": "ID do gênero",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/Genre"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#components/responses/unauthorized"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Genre": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}